feat(0157): носитель числового значения — i128 на всём тракте (задача 0157-01)
u64 и [bit;64] стали выразимы: 0xFFFFFFFFFFFFFFFF и 18446744073709551615
принимаются лексером, доезжают до всех пяти целей и исполняются симулятором.
Носитель — i128 от Token::Number до eval::Value::Number; значение варианта
перечисления переехало вместе с литералом (второй носитель числа завёл бы
границу между ними).

Граница приёма перенесена с носителя на типы языка: [i64::MIN, u64::MAX] —
объединение диапазонов самого широкого знакового и беззнакового типов; шире —
по-прежнему LE-009. Тема выделена в parser/literal_range.rs.

Обёртка u64 заработала: ранний выход bits >= 64 из coerce_integer убран, и
m := m + 1 у границы i64::MAX даёт 9223372036854775808 вместо SIM-004 —
обещание ADR 0127 теперь выполняется и на 64 битах.

Каскад SY-002 снят: литерал вне диапазона даёт одну ошибку, а не две. Адрес
порта сужается явно через SE-055, а не молчаливым as i64.

Печать по типу приёмника (суффикс ULL у цели c, размерная форма у sv) — задачи
0157-03/0157-04; до них критерий A1 не выполнен, и это записано в CHANGES.md.

examples/generated не изменился; precheck.sh зелёный, clippy — ноль
предупреждений. Побочно: lexer.rs вышел за лимит размера, тип LexicalError
выделен в parser/lex_error.rs с сохранением публичного пути.

+++ takt-lang/src/generator/rust/rust_stmt.rs
@@ -671,7 +671,7 @@ mod tail_tests {
     use super::tail_foldable;
     use crate::semantic::{ExpressionNode, StatementNode};
 
-    fn ret(n: i64) -> StatementNode {
+    fn ret(n: i128) -> StatementNode {
         StatementNode::Return(Some(Box::new(ExpressionNode::Number(n))))
     }
     fn if_(then_: StatementNode, else_: Option<StatementNode>) -> StatementNode {
